new messages notifier

+++ ROCKET-BOT/src/main/java/mavmi/telegram_bot/rocket_bot/telegram_bot/Bot.java
@@ -5,20 +5,23 @@ import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.UpdatesListener;
 import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.Update;
+import com.pengrad.telegrambot.model.request.ParseMode;
 import com.pengrad.telegrambot.request.SendMessage;
 import mavmi.telegram_bot.common.auth.BotNames;
 import mavmi.telegram_bot.common.auth.UserAuthentication;
 import mavmi.telegram_bot.common.bot.AbsTelegramBot;
+import mavmi.telegram_bot.common.database.model.RocketImModel;
 import mavmi.telegram_bot.common.database.model.RocketUserModel;
+import mavmi.telegram_bot.common.database.repository.RocketImRepository;
 import mavmi.telegram_bot.common.database.repository.RocketUserRepository;
 import mavmi.telegram_bot.common.logger.Logger;
 import mavmi.telegram_bot.rocket_bot.httpHandler.HttpHandler;
+import mavmi.telegram_bot.rocket_bot.jsonHandler.model.ImHistoryResponse;
+import mavmi.telegram_bot.rocket_bot.jsonHandler.model.ImListResponse;
 import mavmi.telegram_bot.rocket_bot.jsonHandler.model.LoginResponse;
 import mavmi.telegram_bot.rocket_bot.jsonHandler.model.MeResponse;
 
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
 import static mavmi.telegram_bot.rocket_bot.constants.Levels.*;
 import static mavmi.telegram_bot.rocket_bot.constants.Phrases.*;
@@ -26,20 +29,35 @@ import static mavmi.telegram_bot.rocket_bot.constants.Requests.*;
 
 public class Bot extends AbsTelegramBot {
     private final TelegramBot telegramBot;
+    private final Long sleepTime;
     private final UserAuthentication userAuthentication;
     private final RocketUserRepository rocketUserRepository;
+    private final RocketImRepository rocketImRepository;
     private final HttpHandler httpHandler;
     private final Long adminId;
     private final Map<Long, LocalUser> localUsers;
+    private final Notifier notifier;
 
-    public Bot(String telegramBotToken, UserAuthentication userAuthentication, RocketUserRepository rocketUserRepository, HttpHandler httpHandler, Logger logger, Long adminId) {
+    public Bot(
+            String telegramBotToken,
+            Long sleepTime,
+            UserAuthentication userAuthentication,
+            RocketUserRepository rocketUserRepository,
+            RocketImRepository rocketImRepository,
+            HttpHandler httpHandler,
+            Logger logger,
+            Long adminId
+    ) {
         super(logger);
         this.telegramBot = new TelegramBot(telegramBotToken);
+        this.sleepTime = sleepTime;
         this.userAuthentication = userAuthentication;
         this.rocketUserRepository = rocketUserRepository;
+        this.rocketImRepository = rocketImRepository;
         this.httpHandler = httpHandler;
         this.adminId = adminId;
-        localUsers = new HashMap<>();
+        this.localUsers = new HashMap<>();
+        this.notifier = startThread();
     }
 
     @Override
@@ -54,17 +72,18 @@ public class Bot extends AbsTelegramBot {
 
                 if (!userAuthentication.isPrivilegeGranted(userid, BotNames.ROCKET_BOT)) continue;
                 if (inputText == null) continue;
-                logEvent(message);
                 LocalUser localUser = processLocalUser(chatId);
                 int userMenuLevel = localUser.getMenuLevel();
 
                 if (userMenuLevel == MAIN_MENU_LEVEL) {
+                    if (!inputText.equals(FEEDBACK_REQ)) logEvent(message);
                     switch (inputText) {
                         case START_REQ -> start(localUser);
                         case LOGIN_REQ -> login(localUser, inputText);
                         case LOGOUT_REQ -> logout(localUser);
                         case ME_REQ -> me(localUser);
                         case FEEDBACK_REQ -> feedback(localUser, inputText);
+                        case SHOW_CONTENT_REQ -> showContent(localUser);
                     }
                 } else {
                     if (inputText.equals(CANCEL_REQ)) {
@@ -86,31 +105,97 @@ public class Bot extends AbsTelegramBot {
         });
     }
 
-    synchronized void sendMsg(SendMessage sendMessage){
-        telegramBot.execute(sendMessage);
+    synchronized void sendMsg(long chatId, String message, ParseMode parseMode) {
+        telegramBot.execute(new SendMessage(chatId, message).parseMode(parseMode));
+    }
+    synchronized void sendMsg(long chatId, String message) {
+        sendMsg(chatId, message, ParseMode.Markdown);
+    }
+    synchronized void updateLastMessages(long chatId, boolean notifyUsername) {
+        if (!preExecutionTokenCheckout(chatId)) {
+            return;
+        }
+
+        RocketUserModel rocketUserModel = rocketUserRepository.get(chatId);
+        String rcUid = rocketUserModel.getRc_uid();
+        String rcToken = rocketUserModel.getRc_token();
+
+        Map<String, String> chatIdToKnowMsgId = new HashMap<>();
+        for (RocketImModel rocketImModel : rocketImRepository.get(rcUid)) {
+            chatIdToKnowMsgId.put(rocketImModel.getChat_id(), rocketImModel.getLast_msg_id());
+        }
+
+        List<ImListResponse> newMessages = httpHandler.imList(rcUid, rcToken, chatIdToKnowMsgId);
+        if (newMessages == null) {
+            return;
+        }
+
+        for (ImListResponse newMessage : newMessages) {
+            if (notifyUsername) {
+                List<ImHistoryResponse> historyResponses = newMessage.getHistoryResponses();
+                if (rocketUserModel.getShow_content()) {
+                    for (ImHistoryResponse unreadMessage : historyResponses) {
+                        String msg = "***" +
+                                " > " +
+                                unreadMessage.getAuthor_name() +
+                                " [" +
+                                unreadMessage.getTimestamp() +
+                                "]" +
+                                ": " +
+                                "***" +
+                                "\n" +
+                                unreadMessage.getMsg();
+                        sendMsg(chatId, msg);
+                    }
+                } else {
+                    if (!historyResponses.isEmpty()) {
+                        sendMsg(
+                                chatId,
+                                newMessage.getHistoryResponses().size() +
+                                        " новых сообщение от " +
+                                        newMessage.getLast_msg_author_name()
+                        );
+                    }
+                }
+            }
+
+            rocketImRepository.add(
+                    RocketImModel.builder()
+                            .rc_uid(newMessage.getRc_uid())
+                            .chat_id(newMessage.getChat_id())
+                            .last_msg_id(newMessage.getLast_msg_id())
+                            .last_msg_author_id(newMessage.getLast_msg_author_id())
+                            .build()
+            );
+        }
     }
 
     private void start(LocalUser localUser){
-        sendMsg(new SendMessage(localUser.getChatId(), GREETINGS_MSG));
+        sendMsg(localUser.getChatId(), GREETINGS_MSG);
     }
     private void login(LocalUser localUser, String inputText) {
         int menuLevel = localUser.getMenuLevel();
         long chatId = localUser.getChatId();
 
         if (menuLevel == MAIN_MENU_LEVEL) {
-            sendMsg(new SendMessage(chatId, ENTER_LOGIN_DATA_MSG));
+            sendMsg(chatId, ENTER_LOGIN_DATA_MSG);
             localUser.setMenuLevel(ENTER_LOGIN_DATA_LEVEL);
         } else if (menuLevel == ENTER_LOGIN_DATA_LEVEL) {
             String[] loginPasswd = inputText.split("\n");
             if (loginPasswd.length != 2) {
-                sendMsg(new SendMessage(chatId, INVALID_DATA_FORMAT_MSG));
+                sendMsg(chatId, INVALID_DATA_FORMAT_MSG);
             } else {
                 String username = loginPasswd[0];
                 String passwd = loginPasswd[1];
-                if (auth(username, passwd, chatId)) {
-                    sendMsg(new SendMessage(chatId, LOGIN_SUCCESS_MSG));
+
+                if (rocketUserRepository.get(username) != null) {
+                    sendMsg(chatId, USER_DUPLICATE_MSG);
+                } else if (auth(username, passwd, chatId)) {
+                    updateLastMessages(chatId, false);
+                    notifier.addUser(chatId);
+                    sendMsg(chatId, LOGIN_SUCCESS_MSG);
                 } else {
-                    sendMsg(new SendMessage(chatId, LOGIN_FAIL_MSG));
+                    sendMsg(chatId, LOGIN_FAIL_MSG);
                 }
             }
             localUser.setMenuLevel(MAIN_MENU_LEVEL);
@@ -118,19 +203,34 @@ public class Bot extends AbsTelegramBot {
     }
     private void logout(LocalUser localUser) {
         long chatId = localUser.getChatId();
+        RocketUserModel rocketUserModel = getRocketUserById(chatId);
 
-        rocketUserRepository.delete(chatId);
-        sendMsg(new SendMessage(chatId, LOGOUT_SUCCESS_MSG));
+        if (rocketUserModel != null) {
+            rocketImRepository.delete(rocketUserModel.getRc_uid());
+            rocketUserRepository.delete(chatId);
+            notifier.deleteUser(chatId);
+
+            sendMsg(chatId, LOGOUT_SUCCESS_MSG);
+        }
     }
     private void me(LocalUser localUser) {
         long chatId = localUser.getChatId();
 
-        if (!preExecutionCheckout(chatId)) {
-            sendMsg(new SendMessage(chatId, EXECUTION_FAIL_MSG));
+        if (!preExecutionTokenCheckout(chatId)) {
+            sendMsg(chatId, EXECUTION_FAIL_MSG);
         } else {
             RocketUserModel rocketUserModel = getRocketUserById(chatId);
             MeResponse meResponse = httpHandler.me(rocketUserModel.getRc_uid(), rocketUserModel.getRc_token());
-            sendMsg(new SendMessage(chatId, meResponse.toString()));
+            if (meResponse != null) {
+                sendMsg(
+                        chatId,
+                "***Username: ***" + meResponse.getUsername() + "\n" +
+                        "***Email: ***" + meResponse.getEmail() + "\n" +
+                        "***Имя: ***" + meResponse.getName() + "\n" +
+                        "***Статус профиля: ***" + meResponse.getStatusText() + "\n" +
+                        "***Статус подключения: ***" + meResponse.getStatusConnection()
+                );
+            }
         }
     }
     private void feedback(LocalUser localUser, String inputText) {
@@ -138,32 +238,36 @@ public class Bot extends AbsTelegramBot {
         long chatId = localUser.getChatId();
 
         if (menuLevel == MAIN_MENU_LEVEL) {
-            sendMsg(new SendMessage(chatId, ENTER_FEEDBACK_MSG));
+            sendMsg(chatId, ENTER_FEEDBACK_MSG);
             localUser.setMenuLevel(ENTER_FEEDBACK_LEVEL);
         } else if (menuLevel == ENTER_FEEDBACK_LEVEL) {
-            sendMsg(new SendMessage(
+            sendMsg(
                     adminId,
                     "Новый комментарий от пользователя:\n\n" + inputText
-            ));
-            sendMsg(new SendMessage(chatId, THX_MSG));
+            );
+            sendMsg(chatId, THX_MSG);
             localUser.setMenuLevel(MAIN_MENU_LEVEL);
+        }
+    }
+    private void showContent(LocalUser localUser) {
+        long chatId = localUser.getChatId();
+        RocketUserModel rocketUserModel = rocketUserRepository.get(chatId);
+
+        if (rocketUserModel != null) {
+            boolean newValue = !rocketUserModel.getShow_content();
+            rocketUserModel.setShow_content(newValue);
+            rocketUserRepository.add(rocketUserModel);
+            sendMsg(
+                    chatId,
+                    (newValue) ? "включено" : "выключено"
+            );
         }
     }
     private void cancel(LocalUser localUser) {
         localUser.setMenuLevel(MAIN_MENU_LEVEL);
-        sendMsg(new SendMessage(localUser.getChatId(), CANCEL_MSG));
+        sendMsg(localUser.getChatId(), CANCEL_MSG);
     }
 
-    private boolean preExecutionCheckout(long userId) {
-        RocketUserModel rocketUserModel = getRocketUserById(userId);
-        if (rocketUserModel == null) {
-            return false;
-        }
-        if (isTokenActive(rocketUserModel.getToken_exp())) {
-            return true;
-        }
-        return auth(rocketUserModel.getLogin(), rocketUserModel.getPasswd(), userId);
-    }
     private RocketUserModel getRocketUserById(long userId) {
         return rocketUserRepository.get(userId);
     }
@@ -175,6 +279,9 @@ public class Bot extends AbsTelegramBot {
     }
     private boolean auth(String username, String passwd, long chatId) {
         LoginResponse loginResponse = httpHandler.auth(username, passwd);
+        if (loginResponse == null) {
+            return false;
+        }
         JsonObject message = loginResponse.getMessage();
         if (message.has("result")) {
             JsonObject result = message.get("result").getAsJsonObject();
@@ -196,6 +303,18 @@ public class Bot extends AbsTelegramBot {
             return false;
         }
     }
+    private boolean preExecutionTokenCheckout(long userId) {
+        RocketUserModel rocketUserModel = getRocketUserById(userId);
+
+        if (rocketUserModel == null) {
+            return false;
+        }
+        if (isTokenActive(rocketUserModel.getToken_exp())) {
+            return true;
+        }
+
+        return auth(rocketUserModel.getLogin(), rocketUserModel.getPasswd(), userId);
+    }
 
     private LocalUser processLocalUser(long chatId){
         LocalUser localUser = localUsers.get(chatId);
@@ -206,5 +325,17 @@ public class Bot extends AbsTelegramBot {
         }
 
         return localUser;
+    }
+    private Notifier startThread() {
+        Set<Long> userIdx = new HashSet<>();
+
+        for (RocketUserModel rocketUserModel : rocketUserRepository.getAll()) {
+            userIdx.add(rocketUserModel.getUserid());
+        }
+
+        Notifier tmpNotifier = new Notifier(this, sleepTime, userIdx);
+        tmpNotifier.start();
+
+        return tmpNotifier;
     }
 }
